Project concluded

+++ app/Http/Controllers/PostsController.php
@@ -3,11 +3,21 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 
 class PostsController extends Controller
 {
+      /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -44,13 +54,31 @@ class PostsController extends Controller
     {
       $this->validate($request, [
         'title' => 'required',
-        'body' => 'required'
+        'body' => 'required',
+        'cover_image' => 'image|nullable|max:1999'
       ]);
-      // Create Post
+
+      // Handle file Upload
+
+      if($request->hasFile('cover_image')){
+        // Get filename with the extionsion
+        $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+        // Get just filename
+        $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+        // Getjust ext
+        $extension = $request->file('cover_image')->getClientOriginalExtension();
+        // Filename to store
+        $fileNameToStore = $filename . '_' . time() . '.' . $extension;
+        $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+      } else {
+        $fileNameToStore = 'noimage.jpg';
+      }
+              // Create Post
       $post = new Post;
       $post->title = $request->input('title');
       $post->body = $request->input('body');
       $post->user_id = auth()->user()->id;
+      $post->cover_image = $fileNameToStore;
       $post->save();
       return redirect('/posts')->with('success', 'Post Created');
     }
@@ -75,7 +103,14 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
+
         $post = Post::find($id);
+
+        //Check for correct user
+        if(auth()->user()->id !== $post->user_id) {
+          return redirect('/posts')->with('error', "Unauthorized Page");
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -92,10 +127,27 @@ class PostsController extends Controller
         'title' => 'required',
         'body' => 'required'
       ]);
+
+      // Handle file Upload
+
+      if($request->hasFile('cover_image')){
+        // Get filename with the extionsion
+        $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+        // Get just filename
+        $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+        // Getjust ext
+        $extension = $request->file('cover_image')->getClientOriginalExtension();
+        // Filename to store
+        $fileNameToStore = $filename . '_' . time() . '.' . $extension;
+        $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+      }
       // Create Post
       $post = Post::find($id);
       $post->title = $request->input('title');
       $post->body = $request->input('body');
+      if($request->hasFile('cover_image')){
+        $post->cover_image = $fileNameToStore;
+      }
       $post->save();
       return redirect('/posts')->with('success', 'Post Updated');
     }
@@ -108,6 +160,15 @@ class PostsController extends Controller
      */
     public function destroy($id)
     {
+        //Check for correct user
+        if(auth()->user()->id !== $post->user_id) {
+          return redirect('/posts')->with('error', "Unauthorized Page");
+        }
+
+        if($post->cover_image != 'noimage.jpg') {
+          // Delete Image
+          Storage::delete('public/cover_images/' . $post->cover_image);
+        }
         $post = Post::find($id);
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
